move basic_github to an examples dir
